+++ frameworks/fmi2/src/main/java/org/intocps/maestro/framework/fmi2/api/mabl/scoping/AMaBLScope.java
@@ -7,8 +7,8 @@ import org.intocps.maestro.ast.node.*;
 import org.intocps.maestro.framework.fmi2.Fmi2SimulationEnvironment;
 import org.intocps.maestro.framework.fmi2.api.Fmi2Builder;
 import org.intocps.maestro.framework.fmi2.api.mabl.AMaBLVariableCreator;
-import org.intocps.maestro.framework.fmi2.api.mabl.AMablValue;
 import org.intocps.maestro.framework.fmi2.api.mabl.MablApiBuilder;
+import org.intocps.maestro.framework.fmi2.api.mabl.values.AMablValue;
 import org.intocps.maestro.framework.fmi2.api.mabl.variables.AMablDoubleVariable;
 import org.intocps.maestro.framework.fmi2.api.mabl.variables.AMablVariable;
 import org.intocps.orchestration.coe.modeldefinition.ModelDescription;
