+++ frameworks/fmi2/src/main/java/org/intocps/maestro/framework/fmi2/api/mabl/values/MablDoubleValue.java
@@ -1,7 +1,6 @@
 package org.intocps.maestro.framework.fmi2.api.mabl.values;
 
 import org.intocps.maestro.framework.fmi2.api.Fmi2Builder;
-import org.intocps.maestro.framework.fmi2.api.mabl.AMablValue;
 
 import static org.intocps.maestro.ast.MableAstFactory.newRealType;
