+++ frameworks/fmi2/src/main/java/org/intocps/maestro/framework/fmi2/api/mabl/variables/AMablFmi2ComponentVariable.java
@@ -2,8 +2,12 @@ package org.intocps.maestro.framework.fmi2.api.mabl.variables;
 
 import org.intocps.maestro.ast.node.*;
 import org.intocps.maestro.framework.fmi2.api.Fmi2Builder;
-import org.intocps.maestro.framework.fmi2.api.mabl.*;
+import org.intocps.maestro.framework.fmi2.api.mabl.AMablPort;
+import org.intocps.maestro.framework.fmi2.api.mabl.MablApiBuilder;
+import org.intocps.maestro.framework.fmi2.api.mabl.ModelDescriptionContext;
+import org.intocps.maestro.framework.fmi2.api.mabl.PortIdentifier;
 import org.intocps.maestro.framework.fmi2.api.mabl.scoping.IMablScope;
+import org.intocps.maestro.framework.fmi2.api.mabl.values.PortValueMapImpl;
 import org.intocps.orchestration.coe.modeldefinition.ModelDescription;
 
 import java.util.*;
